requirements.txt
streamlit
pandas
python-docx
plotly
matplotlib        #  NEW
ibm-watson
ibm-cloud-sdk-core
